agent/requirements.txt
Flask
connexion
Flask-Injector
fastavro
pyzmq
networkx
jsonrpcserver
jsonrpcclient
